+++ controllers/comment.ts
@@ -1,11 +1,14 @@
 import { Request, Response } from "express";
+import { Types } from "mongoose";
 import { CommentService } from "../services";
 
 export default class CommentController {
     static async findById(req: Request, res: Response, next: Function): Promise<any> {
         try {
-            const article = await CommentService.findById(req.params.id);
-            res.send(article);
+            if (!Types.ObjectId.isValid(req.params.id))
+                res.status(400).send({ message: "Invalid id" });
+            else
+                res.send(await CommentService.findById(req.params.id));
         } catch (err) {
             next(err);
         }
@@ -13,7 +16,11 @@ export default class CommentController {
 
     static async fetch(req: Request, res: Response, next: Function): Promise<any> {
         try {
-            res.send(await CommentService.findByArticleId(req.query.articleId as string));
+            if (!Types.ObjectId.isValid(req.params.id))
+                res.status(400).send({ message: "Invalid id" });
+            else
+                res.send(await CommentService.findByArticleId(req.params.id));
+
         } catch (err) {
             next(err);
         }
@@ -21,8 +28,7 @@ export default class CommentController {
 
     static async create(req: Request, res: Response, next: Function): Promise<any> {
         try {
-            const article = await CommentService.create(req.body);
-            res.send(article);
+            res.send(await CommentService.create(req.body));
         }
         catch (err) {
             next(err);
@@ -31,8 +37,12 @@ export default class CommentController {
 
     static async update(req: Request, res: Response, next: Function): Promise<any> {
         try {
-            const article = await CommentService.update(req.params.id, req.body);
-            res.send(article);
+            if (!Types.ObjectId.isValid(req.params.id))
+                res.status(400).send({ message: "Invalid id" });
+            else if (!await CommentService.findById(req.params.id))
+                res.status(404).send({ message: "Comment not found" });
+            else
+                res.send(await CommentService.update(req.params.id, req.body));
         } catch (err) {
             next(err);
         }
@@ -40,8 +50,14 @@ export default class CommentController {
 
     static async remove(req: Request, res: Response, next: Function): Promise<any> {
         try {
-            await CommentService.delete(req.params.id);
-            res.end();
+            if (!Types.ObjectId.isValid(req.params.id))
+                res.status(400).send({ message: "Invalid id" });
+            else if (!await CommentService.findById(req.params.id)){
+                res.status(404).send({ message: "Comment not found" });
+            } else {
+                await CommentService.delete(req.params.id);
+                res.end();
+            }
         } catch (err) {
             next(err);
         }
